Revert "code separation, tests and more details"
This reverts commit f8c7df3ba2cdf56b5a9f6c34cb6c1e050876a24f.
